refactor(theme): optimize module imports
