+++ src/framework/theme/component/mapping/mappingContext.ts
@@ -4,4 +4,6 @@ import { ThemeMappingType } from './type';
 const defaultValue: ThemeMappingType = {};
 const MappingContext = React.createContext(defaultValue);
 
-export default MappingContext;
+export {
+  MappingContext,
+};
