+++ src/framework/theme/component/theme/themeContext.ts
@@ -4,4 +4,6 @@ import { ThemeType } from './type';
 const defaultValue: ThemeType = {};
 const ThemeContext = React.createContext(defaultValue);
 
-export default ThemeContext;
+export {
+  ThemeContext,
+};
