+++ src/framework/theme/component/theme/themeProvider.component.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from 'react';
-import ThemeContext from './themeContext';
+import { ThemeContext } from './themeContext';
 import { ThemeType } from './type';
 
 export interface Props {
